♻️ refactor : 로직 나누기
