+++ src/refactoring/components/admin/coupon/CouponForm.tsx
@@ -1,27 +1,12 @@
-import { useState } from "react"
 import { Coupon } from "../../../../types"
+import { useNewCoupon } from "../../../hooks"
 
 type Props = {
   onCouponAdd: (newCoupon: Coupon) => void
 }
 
 const CouponForm = ({ onCouponAdd }: Props) => {
-  const [newCoupon, setNewCoupon] = useState<Coupon>({
-    name: "",
-    code: "",
-    discountType: "percentage",
-    discountValue: 0,
-  })
-
-  const handleAddCoupon = () => {
-    onCouponAdd(newCoupon)
-    setNewCoupon({
-      name: "",
-      code: "",
-      discountType: "percentage",
-      discountValue: 0,
-    })
-  }
+  const { newCoupon, updateNewCoupon, handleAddCoupon } = useNewCoupon({ onCouponAdd })
 
   return (
     <div className="space-y-2 mb-4">
@@ -29,24 +14,21 @@ const CouponForm = ({ onCouponAdd }: Props) => {
         type="text"
         placeholder="쿠폰 이름"
         value={newCoupon.name}
-        onChange={(e) => setNewCoupon({ ...newCoupon, name: e.target.value })}
+        onChange={(e) => updateNewCoupon("name", e.target.value)}
         className="w-full p-2 border rounded"
       />
       <input
         type="text"
         placeholder="쿠폰 코드"
         value={newCoupon.code}
-        onChange={(e) => setNewCoupon({ ...newCoupon, code: e.target.value })}
+        onChange={(e) => updateNewCoupon("code", e.target.value)}
         className="w-full p-2 border rounded"
       />
       <div className="flex gap-2">
         <select
           value={newCoupon.discountType}
           onChange={(e) =>
-            setNewCoupon({
-              ...newCoupon,
-              discountType: e.target.value as "amount" | "percentage",
-            })
+            updateNewCoupon("discountType", e.target.value as "amount" | "percentage")
           }
           className="w-full p-2 border rounded"
         >
@@ -57,7 +39,7 @@ const CouponForm = ({ onCouponAdd }: Props) => {
           type="number"
           placeholder="할인 값"
           value={newCoupon.discountValue}
-          onChange={(e) => setNewCoupon({ ...newCoupon, discountValue: parseInt(e.target.value) })}
+          onChange={(e) => updateNewCoupon("discountValue", parseInt(e.target.value))}
           className="w-full p-2 border rounded"
         />
       </div>
