+++ src/refactoring/components/admin/product/ProductItemDetail.tsx
@@ -1,44 +1,18 @@
-import { useEffect, useState } from "react"
 import { Product } from "../../../../types"
 import DiscountList from "./DiscountList"
 import UpdateProductItemForm from "./UpdateProductItemForm"
+import { useUpdateProductForm } from "../../../hooks"
 
 type Props = {
   product: Product
   onProductUpdate: (updatedProduct: Product) => void
 }
 
-type FormDataType = {
-  name: string
-  price: number
-  stock: number
-}
-
 const ProductItemDetail = ({ product, onProductUpdate }: Props) => {
-  const [isEdit, setEditMode] = useState<boolean>(false)
-  const [formData, setFormData] = useState<FormDataType>({ name: "", price: 0, stock: 0 })
-
-  useEffect(() => {
-    const { name, price, stock } = product
-    setFormData({ name, price, stock })
-  }, [product])
-
-  //   form 입력 시 업데이트
-  const updateFormData = (key: string, value: string | number) => {
-    setFormData((prev) => ({ ...prev, [key]: value }))
-  }
-
-  //   수정완료
-  const completeModify = () => {
-    const { name, price, stock } = formData
-
-    if (!name || !price || !stock) {
-      return alert("모든 항목을 입력해주세요")
-    }
-
-    onProductUpdate({ ...product, name, price, stock })
-    setEditMode(false)
-  }
+  const { isEdit, formData, updateFormData, completeModify, toggleEditMode } = useUpdateProductForm(
+    product,
+    onProductUpdate
+  )
 
   return (
     <>
@@ -55,7 +29,7 @@ const ProductItemDetail = ({ product, onProductUpdate }: Props) => {
         </button>
       ) : (
         <button
-          onClick={() => setEditMode(true)}
+          onClick={toggleEditMode}
           data-testid="modify-button"
           className="text-white px-2 py-1 rounded bg-blue-500 hover:bg-blue-600 mt-2"
         >
