+++ src/refactoring/hooks/index.ts
@@ -1,7 +1,11 @@
-export * from "./useCart.ts"
-export * from "./useCoupon.ts"
-export * from "./useProduct.ts"
+export * from "./useCart"
+export * from "./useCoupon"
+export * from "./useProduct"
 
 //추가
-export * from "./useToggle.ts"
-export * from "./useNewProductForm.ts"
+export * from "./useToggle"
+export * from "./useNewProductForm"
+export * from "./useNewCoupon"
+export * from "./useAdmin"
+export * from "./useDiscount"
+export * from "./useUpdateProductForm"
