+++ src/refactoring/hooks/useNewProductForm.ts
@@ -1,16 +1,23 @@
 import { useState } from "react"
-import { Discount, Product } from "../../types"
+import { Product } from "../../types"
+import { initProduct } from "../constants/initialData"
 
-export const useNewProductForm = (initProduct: Omit<Product, "id">) => {
+export const useNewProductForm = () => {
   const [newProductForm, setNewProductForm] = useState<Omit<Product, "id">>(initProduct)
 
-  const updateNewProductForm = (
-    key: keyof Omit<Product, "id" | "discounts">,
-    value: string | number | Discount[]
-  ) => {
+  /**
+   * 새로운 상품 업데이트
+   * @param key
+   * @param value
+   */
+  const updateNewProductForm = <K extends keyof Product>(key: K, value: Product[K]) => {
     setNewProductForm((prev) => ({ ...prev, [key]: value }))
   }
 
+  /**
+   * 새로운 상품 폼 초기화
+   * @returns
+   */
   const resetNewProductForm = () => setNewProductForm(initProduct)
 
   return { newProductForm, updateNewProductForm, resetNewProductForm }
